refactor: rename genre to playlist
